// src/types/Aluno.ts
export type Aluno = {
  id: string;
  nome: string;
  email: string;
  cidade: string;
  cep: string;
  estado: string;
  logradouro: string;
  numeroCasa: string;
  profissao: string;
  cpfCnpj: string;
  dataCadastro?: string;
  status?: string;
  modalidadeDeAula?: string;
  numeroContato?: string;
};